+++ indica/configs/profilers/profiler_gauss.yaml
@@ -7,15 +7,29 @@ electron_density:
     wcenter: 0.3
     wped: 6
 
+impurity_density:ar:
+  y0: 1.0e+17
+  y1: 1.0e+15
+  yend: 1.0e+15
+  wcenter: 0.3
+  wped: 3
+  peaking: 3
 
-impurity_density:
-    y0: 5.0e+16
-    y1: 1.0e+15
-    yend: 1.0e+15
-    peaking: 3
-    wcenter: 0.3
-    wped: 6
+impurity_density:c:
+  y0: 1.0e+17
+  y1: 1.0e+16
+  yend: 1.0e+16
+  wcenter: 0.3
+  wped: 3
+  peaking: 3
 
+impurity_density:he:
+  y0: 1.0e+17
+  y1: 1.0e+16
+  yend: 1.0e+16
+  wcenter: 0.3
+  wped: 3
+  peaking: 3
 
 neutral_density:
     y0: 1.0e+13
