+++ indica/workflows/bda/bayes_workflow.py
@@ -12,8 +12,7 @@ from indica.plotters.plot_bda import plot_bayes_result
 from indica.workflows.bda.bayesblackbox import BayesBlackBox
 from indica.workflows.bda.model_coordinator import ModelCoordinator
 from indica.workflows.bda.optimiser_context import EmceeOptimiser
-from indica.workflows.bda.plasma_profiler import map_plasma_profile_to_midplane
-from indica.workflows.bda.plasma_profiler import PlasmaProfiler
+from indica.models.plasma import PlasmaProfiler
 from indica.workflows.bda.priors import PriorManager
 from indica.writers.bda_tree import create_nodes
 from indica.writers.bda_tree import does_tree_exist
@@ -328,8 +327,8 @@ class BayesWorkflow:
             _blob = [result["blobs"][key] for result in results]
             blobs[key] = xr.concat(_blob, self.plasma_profiler.plasma.t)
         self.blobs = blobs
-        self.midplane_blobs = map_plasma_profile_to_midplane(
-            self.plasma_profiler.plasma, blobs
+        self.midplane_blobs = self.plasma_profiler.map_plasma_profile_to_midplane(
+            blobs
         )
 
         opt_samples = {}
