+++ indica/workflows/bda/bda_run.py
@@ -22,7 +22,7 @@ from indica.workflows.bda.optimiser_context import BOSettings
 from indica.workflows.bda.optimiser_context import EmceeOptimiser
 from indica.workflows.bda.optimiser_context import EmceeSettings
 from indica.workflows.bda.pca import pca_workflow
-from indica.workflows.bda.plasma_profiler import PlasmaProfiler
+from indica.models.plasma import PlasmaProfiler
 from indica.workflows.bda.priors import PriorManager
